+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added type names annotations support - check `VersionedTypeNamesAnnotations`
+
 ## [0.8.0] - 2026-04-08
 
 ### Added